requirements.txt
# Elderly Healthcare Management System - Production Requirements

# Core Framework
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# Database (Optional - using JSON for now)
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20

# Authentication & Security
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1

# OCR & Image Processing (Optional)
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.74

# API & External Services (Optional)
requests==2.31.0
google-cloud-vision==3.4.0
twilio==8.10.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Additional Utilities
Jinja2==3.1.2
MarkupSafe==2.1.1
